+++ src/client/java/io/github/tr100000/codec2schema/compat/fabric/FabricCompatClient.java
@@ -3,6 +3,8 @@ package io.github.tr100000.codec2schema.compat.fabric;
 import io.github.tr100000.codec2schema.ModUtils;
 import io.github.tr100000.codec2schema.api.Codec2SchemaPlugin;
 import io.github.tr100000.codec2schema.api.MapCodecHandlerRegistry;
+import net.minecraft.resources.Identifier;
+import org.jspecify.annotations.Nullable;
 
 public class FabricCompatClient implements Codec2SchemaPlugin {
     @Override
@@ -10,5 +12,15 @@ public class FabricCompatClient implements Codec2SchemaPlugin {
         if (ModUtils.hasFabricModelLoadingApi()) {
             MapCodecHandlerRegistry.register(KeyExistsCodecHandler::predicate, KeyExistsCodecHandler::new);
         }
+    }
+
+    @Override
+    public boolean shouldRun() {
+        return ModUtils.hasFabricModelLoadingApi();
+    }
+
+    @Override
+    public @Nullable Identifier getId() {
+        return ModUtils.id("client/fabric_compat");
     }
 }