+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 ### Added
-- `exportedWith` now contains an `exportedBy` field to show which mod exported the given schema
+- `exportedWith` now contains an `exportedBy` field to show the mod and plugin that exported the given schema
+- Plugins can now have their own IDs, and can be individually disabled through the config